rewriting comments of db/scmsources.vim
trying to preserve the intentions mentioned by ZyX but making it a lot
easeir for users to understand backgrounds and get the key points more
quickly.

Policies such as "4.7. No trailing comma." are superfluous because its
no valid VimL anyway so no need to bother anybody with talking about it.

The https (for github) and .git policies will be found in VAM soon -
again no need to make users read comments. Be lazy and show those
comments if things are done wrong.